+++ packages/iov-bcp/src/atomicswapmerger.ts
@@ -1,9 +1,9 @@
 import { Encoding } from "@iov/encoding";
 
 import { AtomicSwap, OpenSwap, SwapState } from "./atomicswaptypes";
-import { SwapClaimTransaction, SwapTimeoutTransaction } from "./transactions";
+import { SwapAbortTransaction, SwapClaimTransaction } from "./transactions";
 
-function settleAtomicSwap(swap: OpenSwap, tx: SwapClaimTransaction | SwapTimeoutTransaction): AtomicSwap {
+function settleAtomicSwap(swap: OpenSwap, tx: SwapClaimTransaction | SwapAbortTransaction): AtomicSwap {
   if (tx.kind === "bcp/swap_claim") {
     return {
       kind: SwapState.Claimed,
@@ -20,13 +20,13 @@ function settleAtomicSwap(swap: OpenSwap, tx: SwapClaimTransaction | SwapTimeout
 
 export class AtomicSwapMerger {
   private readonly open = new Map<string, OpenSwap>();
-  private readonly settling = new Map<string, SwapClaimTransaction | SwapTimeoutTransaction>();
+  private readonly settling = new Map<string, SwapClaimTransaction | SwapAbortTransaction>();
 
   /**
    * Takes an event, checks if there is already a matching open or settling event
    * stored in the pool and merges.
    */
-  public process(event: OpenSwap | SwapClaimTransaction | SwapTimeoutTransaction): AtomicSwap | undefined {
+  public process(event: OpenSwap | SwapClaimTransaction | SwapAbortTransaction): AtomicSwap | undefined {
     switch (event.kind) {
       case SwapState.Open: {
         const idAsHex = Encoding.toHex(event.data.id);